Fix type mismatch

+++ 2025/Day03.cs
@@ -44,14 +44,14 @@ public static class DayThreeProgram
 
     foreach (var line in lines)
     {
-      var pack = line.Select(d => (long)d - (long)'0').ToArray();
+      var pack = line.Select(d => (int)d - (int)'0').ToArray();
       var packJoltage = 0L;
       var indexes = new int[12];
       indexes[0] = PosOfMax(pack, 0, pack.Length - indexes.Length);
       for (int i = 1; i < indexes.Length; i++)
       {
         indexes[i] = PosOfMax(pack, indexes[i-1] + 1, pack.Length - indexes.Length + i);
-        packJoltage *= 10;
+        packJoltage *= 10L;
         packJoltage += pack[indexes[i]];
       }
